requirements.txt
astronomer-cosmos
apache-airflow-providers-google
pyarrow
numpy
pandas
protobuf
Requests
sodapy
pendulum


